Add simple README with licence and copyright info
As a third party whose attention just drifted towards the repo, I'd like a little README just noting what this data is, where it comes from and how it's licensed. It's fairly easy to find this out but adding a README for a future @rjw57 is a small contribution I could make. (Writing documentation is always good, yes?)

I've tried to add a small factual README which is fairy uncontroversial but obviously I didn't create this repo or generate the data so do feel very free to correct/amend/delete as appropriate.

Finally, I don't think one is *compelled* to include a copy of the OS OpenData licence in derivative works such as this but it feels like one is being a good citizen by at least linking to it so I added that information to.